refactor: fix missing imports (part 2)
