+++ src/renderer/app/models/tab.ts
@@ -3,12 +3,6 @@ import * as React from 'react';
 import { ipcRenderer } from 'electron';
 
 import store from '~/renderer/app/store';
-import {
-  TABS_PADDING,
-  defaultTabOptions,
-  TAB_ANIMATION_DURATION,
-  ADD_TAB_MARGIN_LEFT,
-} from '~/renderer/app/constants';
 import { ISite } from '~/interfaces';
 
 let id = 1;
